requirements.txt
langgraph>=0.2.0
openai>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
pytest>=8.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
